# specs/002-multitenant-tokenization/contracts/api.yaml
openapi: 3.1.0
info:
  title: PCI Token Vault — Multitenant Extensions
  version: 2.0.0
  description: >
    New and modified endpoints for multitenant support.
    All endpoints (except /health and /admin/*) now require
    an X-Tenant-ID header.

# ─── New: Tenant identification ──────────────────────────────
components:
...
  schemas:
...
    TenantResponse:
      type: object
      required: [tenant_id, name, status, created_at]
      properties:
        tenant_id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [active, inactive]
        kms_key_arn:
          type: string
          description: Dedicated KMS key ARN (only visible to admin)
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
paths:
...
  /admin/tenants/{tenant_id}:
    get:
      operationId: getTenant
      summary: Get tenant details
      tags: [Tenant Management]
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tenant details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantResponse'
        '404':
          description: Tenant not found

    delete:
      operationId: deactivateTenant
      summary: Deactivate a tenant
      description: >
        Sets tenant status to inactive. All operations for this
        tenant are blocked. Data is preserved for audit.
        Reversible — tenant can be re-activated.
      tags: [Tenant Management]
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tenant deactivated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantResponse'
        '404':
          description: Tenant not found

# ─── Modified endpoints (header addition) ────────────────────
# All existing endpoints now require X-Tenant-ID header:
#   POST /vault/tokenize        → scoped by tenant
#   POST /internal/detokenize   → validates token belongs to tenant
#   GET  /vault/tokens/{id}     → scoped by tenant
#   DELETE /vault/tokens/{id}   → scoped by tenant
#   GET  /vault/tokens/{id}/audit → scoped by tenant
#   POST /proxy/forward         → tenant header forwarded to Tokenizer